refactor(postgrest): move typed builders to SupabaseSwiftMacros; fix swift-syntax range
- Move TypedBuilders/ and PostgrestClient+Typed.swift from PostgREST to
  SupabaseSwiftMacros so swift-syntax is only compiled when the typed API
  is explicitly imported — PostgREST and Supabase no longer depend on it
- SupabaseSwiftMacros gains PostgREST as a dependency
- PostgrestClient+Typed now uses public schema()/from() API instead of
  constructing PostgrestQueryBuilder internally
- Swift-syntax range widened to 510.0.0..<605.0.0 to match the package's
  minimum Swift 5.10 support (was 600.0.0..<605.0.0)
