+++ Package.swift
@@ -39,8 +39,9 @@ let package = Package(
     .package(url: "https://github.com/WeTransfer/Mocker", from: "3.0.0"),
     .package(url: "https://github.com/21-DOT-DEV/swift-secp256k1.git", from: "0.23.2"),
     .package(url: "https://github.com/krzyzanowskim/CryptoSwift.git", from: "1.10.0"),
-    // swift-macro-testing 0.6.5 accepts swift-syntax "509.0.0"..<"605.0.0", so 600.x is safe.
-    .package(url: "https://github.com/swiftlang/swift-syntax", "600.0.0"..<"605.0.0"),
+    // Lower bound matches the package's minimum Swift version (5.10 → 510.x).
+    // Upper bound matches swift-macro-testing 0.6.x compatibility ceiling.
+    .package(url: "https://github.com/swiftlang/swift-syntax", "510.0.0"..<"605.0.0"),
     .package(url: "https://github.com/pointfreeco/swift-macro-testing", from: "0.6.0"),
   ],
   targets: [
@@ -132,7 +133,6 @@ let package = Package(
         .product(name: "ConcurrencyExtras", package: "swift-concurrency-extras"),
         .product(name: "HTTPTypes", package: "swift-http-types"),
         "Helpers",
-        .target(name: "SupabaseSwiftMacros"),
       ]
     ),
     .testTarget(
@@ -142,6 +142,7 @@ let package = Package(
         "Helpers",
         "Mocker",
         "PostgREST",
+        "SupabaseSwiftMacros",
         "TestHelpers",
       ],
       exclude: [
@@ -214,7 +215,6 @@ let package = Package(
         "PostgREST",
         "Realtime",
         "Storage",
-        .target(name: "SupabaseSwiftMacros"),
       ]
     ),
     .testTarget(
@@ -256,7 +256,8 @@ let package = Package(
     .target(
       name: "SupabaseSwiftMacros",
       dependencies: [
-        .target(name: "SupabaseMacros")
+        .target(name: "SupabaseMacros"),
+        "PostgREST",
       ]
     ),
     .testTarget(
