+++ Sources/SupabaseSwiftMacros/TypedBuilders/TypedPostgrestFilterBuilder.swift
@@ -1,12 +1,12 @@
 //
 //  TypedPostgrestFilterBuilder.swift
-//  PostgREST
+//  SupabaseSwiftMacros
 //
 //  Created by Guilherme Souza on 24/06/25.
 //
 
 import Foundation
-import SupabaseSwiftMacros
+import PostgREST
 
 /// Wraps PostgrestFilterBuilder with typed KeyPath-based filter methods.
 /// Table is constrained to ReadOnlyTableRepresentable (the shared base) so this builder
